+++ UIKitSwagger/View/UIViewSubscripting.swift
@@ -1,5 +1,5 @@
 //
-//  UIViewHierarchy.swift
+//  UIViewSubscripting.swift
 //  UIKitSwagger
 //
 //  Created by Sam Odom on 6/23/14.
@@ -60,24 +60,6 @@ extension UIView {
             }
         }
 
-    }
-
-    /**
-      Searches for the first common view at or above this view and the provided view.
-
-      :param:       view View whose hierarchy should be searched along with the current view.
-      :returns:     First common ancestor of the current and provided views, if they share one.
-    */
-    public func firstCommonAncestor(view: UIView) -> UIView? {
-        if isDescendantOfView(view) {
-            return view
-        }
-
-        if view.isDescendantOfView(self) {
-            return self
-        }
-
-        return superview?.firstCommonAncestor(view)
     }
 
 }