+++ CHANGELOG.md
@@ -8,7 +8,13 @@ UIKitSwagger Changelog
  - Constraint searches no longer return optional arrays of constraints.  Missing constraints will now yield an empty array.
  - The `Constraint` typealias has been removed.
  - Making a positive constant constraint from one that is already non-negative will return the original constraint.
+ - Layout guides are now supported in the constraint builder syntax.
  - Layout guides are now addable/removable to/from views using the `+=` and `-=` syntax.  They are added and removed along with views to make sure that modifier items make sense.
+ - Removed `NSLayoutConstraint.apply()`, `NSLayoutConstraint.remove()`, `ApplyConstraints()` and `RemoveConstraints`.  Use the activation/deactivation functions instead.
+ - Layout constraints are no longer addable/removable to/from views using the `+=` and `-=` syntax.  Use constraint activation/deactivation instead.
+ - Constraint-producing functions that added their constraints to the appropriate view now use activation to accomplish that task.
+ - Removed `firstCommonAncestor` from UIView since constraint activation obviates the need for knowing where to apply a constraint.
+ - Fixed a newly discovered constraint search bug.
 
 
 ##### v0.6.1 (Sunday, March 22, 2015)